just use one loader
